Activities moved from ui to presentation
