+++ app/src/main/java/com/example/playlistmaker/presentation/search/TrackAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui.search
+package com.example.playlistmaker.presentation.search
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
