+++ app/src/main/res/layout/activity_settings.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.settings.SettingsActivity">
+    tools:context=".presentation.settings.SettingsActivity">
 
 
     <LinearLayout
